requirements.txt
pyserial==3.5
matplotlib==3.5.1
pandas==1.4.2
numpy==1.24.0
PyQt5==5.15.0 